+++ src/components/custom-field-input-single-choice/custom-field-input-single-choice.jsx
@@ -1,6 +1,5 @@
 import React, {
   forwardRef,
-  useEffect,
   useImperativeHandle,
   useRef,
   useState,
@@ -15,9 +14,9 @@ import styles from './custom-field-input-single-choice.css';
 import Listbox from '../listbox/listbox.jsx';
 import ListOption from '../list-option/list-option.jsx';
 import useValidation from '../../hooks/use-validation.jsx';
+import useDropdownClose from '../../hooks/use-dropdown-close.js';
 
 const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleChoice(props, ref) {
-  const [didMount, setDidMount] = useState(false);
   const [showOptions, setShowOptions] = useState(false);
   const [value, setValue] = useState(props.value);
   const [searchValue, setSearchValue] = useState(undefined);
@@ -30,6 +29,15 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
     name={props.readOnly ? iconCaretDownDisabled.id : iconCaretDown.id}
     fill="skip"
   />);
+
+  const defaultValue = value ? value.label : '';
+
+  const wrapperRef = useRef(null);
+  const handleDropdownClose = () => {
+    setShowOptions(false);
+    setSearchValue(defaultValue);
+  };
+  useDropdownClose(wrapperRef, showOptions, handleDropdownClose);
 
   const clear = () => {
     setValue(undefined);
@@ -102,6 +110,7 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
     setValue(selectedValue);
     setSearchValue(selectedValue.label);
     setShowOptions(false);
+    inputRef.current.focus();
   }
 
   function onSearchChange(event) {
@@ -117,15 +126,6 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
     setShowOptions(true);
   }
 
-  useEffect(() => {
-    setDidMount(true);
-  }, []);
-
-  useEffect(() => {
-    if (!didMount) return;
-    if (!showOptions) inputRef.current.focus();
-  }, [showOptions]);
-
   useImperativeHandle(ref, () => ({
     id: props.id,
     get value() {
@@ -134,7 +134,7 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
   }));
 
   return (
-    <div className={styles.container}>
+    <div ref={wrapperRef} className={styles.container}>
       <AbstractCustomField
         icon={caretIcon}
         clear={clearIcon()}
@@ -148,7 +148,7 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
         inputRef={inputRef}
         required={props.required}
         errorText={validationMessage}
-        value={searchValue || (value ? value.label : '')}
+        value={searchValue || defaultValue}
       />
       { showOptions && (
         <Listbox