+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
   - Major: All custom fields now use a single `errorText` prop to determine error state
+  - Minor: Update `CustomFieldInputMultipleChoice` and `CustomFieldInputMultipleChoice` to close dropdown on click, escape and tab
 </details>
 
 ## 0.30.0 (July 16, 2020)