+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
-  - Major: All custom fields now use a single `errorText` prop to determine error state
-  - Minor: Update `CustomFieldInputMultipleChoice` and `CustomFieldInputMultipleChoice` to close dropdown on click, escape and tab
 </details>
+
+## 0.31.0 (July 20, 2020)
+- Major: All custom fields now use a single `errorText` prop to determine error state
+- Minor: Update `CustomFieldInputMultipleChoice` and `CustomFieldInputMultipleChoice` to close dropdown on click, escape and tab
 
 ## 0.30.0 (July 16, 2020)
 - Minor: Update `CustomFieldInputSingleChoice` to allow removing selected choices